pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whatnut"
version = "0.2.0"
description = "Monte Carlo analysis of life expectancy from nut consumption"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Max Ghenis", email = "max@maxghenis.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pyyaml>=6.0",
    "pypdf>=6.10.2",
    "openpyxl>=3.1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
docs = [
    "ipykernel>=6.0",
    "jupyter>=1.0",
]
all = [
    "whatnut[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/MaxGhenis/whatnut"
Documentation = "https://maxghenis.github.io/whatnut"

[tool.hatch.build.targets.wheel]
packages = ["src/whatnut"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=whatnut --cov-report=term-missing"

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
